Validate existing tokens & better handling of responses for token requests.
After requesting a new token at startup, when parsing the response, the response codes that are outside of the 200 range result in an invalid token which will not be saved.
Given the fact that
1 - whenever a request for a new token is made, the post data includes the existing token
2 - the api returns the existing token without verification
3 - existing tokens might have been corrupted, therefore the invalid value is perpetuated ad infinitum after a single failure
Then the existing token is validated once and only once, prior to sending it as part of the post data.
If the validation fails, the existing token is cleared. (set to null)